requirements.txt
paho-mqtt>=2.0.0
django==5.2.1
djangorestframework>=3.15
gunicorn~=23.0.0